requirements.txt
langchain
langchain-community
langchain-experimental
langchain-groq
chromadb
sentence-transformers
python-dotenv
pymysql
